// src/types/news.ts
export type News = {
  id: string;
  publishedAt: Date;
  feed: string;
  url: string;
  postType: string;
  block: string;
  section: string;
  title: string;
  desc: string;
  showDesc: boolean;
  mainImage: string;
  mainImgDesc: string;
  mainImgAuthor: string;
  imageSize: boolean;
  content: string;
  live: boolean;
  showAuthorDesc: boolean;
  showAuthor: boolean;
  publishOnSocialMedia: boolean;
  views: number;
  createdAt: Date;
  updatedAt: Date;
  UserId: string;
  userId: string;

  author: string;
};